backend/requirements.txt
# FastAPI and Server
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Pydantic
pydantic==2.5.0
pydantic-settings==2.1.0

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Database
sqlalchemy==2.0.23

# HTTP
aiohttp==3.9.1
httpx==0.25.2
requests==2.31.0

# Data Processing
pandas==2.1.4
numpy==1.26.2

# Market Data
yfinance==0.2.33
feedparser==6.0.10

# Pathway - Use latest version
pathway==0.13.0

# Document Processing
PyPDF2==3.0.1
python-magic==0.4.27

# Google Gemini
google-generativeai==0.3.2

# Utilities
python-dotenv==1.0.0
